test-langchain/requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain==0.0.219
faiss-cpu
tiktoken
streamlit
chromadb==0.3.25
Flask